로그아웃 및 자동 로그인 +Bonjour 관련 코드들 모두 삭제

+++ app/src/main/java/com/example/bohemeow/UserData.java
@@ -9,9 +9,9 @@ public class UserData {
     public String password;
     public int weight;
     public double totalWalkLength;
-    public double totalWalkTime;
+    public long totalWalkTime;
     public int totalWalkCount;
-    public double realWalkTime;
+    public long realWalkTime;
     public int level;
     public int catType;
     public boolean isItemExist;
